Move and add API client-side to use GITHUB_USER
