+++ cli/pyproject.toml
@@ -17,6 +17,7 @@ readme = "README.md"
 description = "Client-sided API interaction installed to Codespaces container."
 
 [project.scripts]
+presence = "omnipresence:main.report"
 inventory = "inventory:main.list"
 get = "inventory:Acquire.cmd"
 use = "inventory:Use.cmd"
