+++ cli/src/inventory/Drop.py
@@ -6,7 +6,6 @@ import requests
 import importlib
 
 from dotenv import load_dotenv
-from getpass import getuser
 from .Instance import Instance
 
 load_dotenv()
@@ -26,7 +25,7 @@ class Dropped:
             f"{os.getenv('API_URL')}:{os.getenv('API_PORT')}/v1/inventory/reduce/",
             data = {
                 "item_name": item_name,
-                "item_owner": getuser(),
+                "item_owner": os.getenv('GITHUB_USER'),
                 "item_drop": True
             }
         )
@@ -35,7 +34,7 @@ class Dropped:
         item = requests.post(
             f"{os.getenv('API_URL')}:{os.getenv('API_PORT')}/v1/inventory/search/",
             data = {
-                "charname": getuser(),
+                "charname": os.getenv('GITHUB_USER'),
                 "item_name": item_name
             }
         )
